% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{glfeat}
\alias{glfeat}
\title{Get a feature entry by key}
\usage{
glfeat(key, default = NULL, error = FALSE)
}
\arguments{
\item{key}{key under \verb{feature/}, e.g. "homelang"}

\item{default}{value returned when key is missing (default NULL)}

\item{error}{if TRUE, throw an error when key is missing}
}
\value{
the resolved value (can be list: e.g. var + level)
}
\description{
Get a feature entry by key
}
