% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{glstat}
\alias{glstat}
\title{Get a statistical term by key}
\usage{
glstat(key, default = NULL, error = FALSE)
}
\arguments{
\item{key}{key under \verb{stat/}, e.g. "p_value"}

\item{default}{value returned when key is missing (default NULL)}

\item{error}{if TRUE, throw an error when key is missing}
}
\value{
the resolved value (usually character)
}
\description{
Get a statistical term by key
}
