% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{glterm}
\alias{glterm}
\title{Get a generic term by key}
\usage{
glterm(key, default = NULL, error = FALSE)
}
\arguments{
\item{key}{key under \verb{term/}, e.g. "and"}

\item{default}{value returned when key is missing (default NULL)}

\item{error}{if TRUE, throw an error when key is missing}
}
\value{
the resolved value (usually character)
}
\description{
Get a generic term by key
}
