% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-objects.R
\docType{data}
\name{JSON}
\alias{JSON}
\title{Active glossary JSON (namespace-scoped)}
\format{
An object of class \code{list} of length 15.
}
\usage{
JSON
}
\description{
This object is managed by ICERglossary and updated by \code{glreset()}, \code{gllang()},
and \code{glread()}. Do not modify it directly.
}
\keyword{datasets}
