% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{glfmt}
\alias{glfmt}
\title{Format values using a glossary format string}
\usage{
glfmt(key, ...)
}
\arguments{
\item{key}{key under \verb{format/}, e.g. "points"}

\item{...}{values to insert; unnamed values are matched by placeholder order}
}
\value{
formatted character string
}
\description{
Format values using a glossary format string
}
